+++ R/PomaUnivariate.R
@@ -13,11 +13,11 @@
 #'
 #' @export
 #'
-#' @return A data frame with results.
+#' @return A tibble with results.
 #' @author Pol Castellano-Escuder
 #'
 #' @importFrom tibble column_to_rownames rownames_to_column remove_rownames
-#' @importFrom dplyr select mutate mutate_all filter bind_cols bind_rows summarise_all group_by as_tibble everything
+#' @importFrom dplyr select mutate mutate_all rename_at filter bind_cols bind_rows summarise_all group_by as_tibble everything rename_all across starts_with vars
 #' @importFrom magrittr %>%
 #' @importFrom SummarizedExperiment assay colData
 #' 
@@ -69,7 +69,7 @@ PomaUnivariate <- function(data,
   Group <- as.factor(SummarizedExperiment::colData(data)[,1])
   e <- t(SummarizedExperiment::assay(data))
 
-  ## group means
+  ## group means and sd
   
   group_means <- e %>%
     as.data.frame() %>% 
@@ -79,9 +79,19 @@ PomaUnivariate <- function(data,
     remove_rownames() %>%
     column_to_rownames("group") %>%
     t() %>%
-    as.data.frame()
+    as.data.frame() %>% 
+    rename_all(~ paste0("mean_", .))
   
-  colnames(group_means) <- paste0("mean_", colnames(group_means))
+  group_sd <- e %>%
+    as.data.frame() %>% 
+    mutate(group = Group) %>%
+    group_by(group) %>%
+    summarise_all(list(~ sd(., na.rm = TRUE))) %>%
+    remove_rownames() %>%
+    column_to_rownames("group") %>%
+    t() %>%
+    as.data.frame() %>% 
+    rename_all(~ paste0("sd_", .))
   
   ##
   
@@ -93,7 +103,7 @@ PomaUnivariate <- function(data,
     res_ttest <- data.frame(pvalue = apply(FUN = stat_ttest, MARGIN = 2, X = e)) %>% 
       rownames_to_column("feature") %>%
       mutate(pvalueAdj = p.adjust(pvalue, method = adjust)) %>%
-      bind_cols(group_means) %>%
+      bind_cols(group_means, group_sd) %>%
       mutate(FC = as.numeric(round(group_means[,2]/group_means[,1], 3)),
              diff_means = as.numeric(round(group_means[,2] - group_means[,1], 3))) %>%
       dplyr::select(feature, FC, diff_means, pvalue, pvalueAdj, everything()) %>% 
@@ -110,7 +120,7 @@ PomaUnivariate <- function(data,
       
       res_aov <- data.frame(pvalue = apply(FUN = stat_aov, MARGIN = 2, X = e)) %>%
         mutate(pvalueAdj = p.adjust(pvalue, method = adjust)) %>% 
-        bind_cols(group_means) %>% 
+        bind_cols(group_means, group_sd) %>% 
         rownames_to_column("feature") %>% 
         dplyr::select(feature, pvalue, pvalueAdj, everything()) %>% 
         dplyr::as_tibble()
@@ -139,16 +149,20 @@ PomaUnivariate <- function(data,
       
       result_cov <- vector(mode = "list", length = ncol(e))
       for(i in 1:ncol(e)) {
-        result_cov[[i]] <- data.frame(pvalue = anova(aov(as.formula(paste(colnames(covariates_feat)[i], "~", 
-                                                                          model_names)),
-                                            data = covariates_feat))$"Pr(>F)"[ncol(covariates)+1])
+        result_cov[[i]] <- anova(aov(as.formula(paste(colnames(covariates_feat)[i], "~", 
+                                                      model_names)),
+                                     data = covariates_feat))$"Pr(>F)"[1:(ncol(covariates)+1)]
+        
+        names(result_cov[[i]]) <- c(colnames(covariates), colnames(SummarizedExperiment::colData(data))[1])
       }
 
       res_aov_cov <- bind_rows(result_cov) %>%
-        mutate(feature = colnames(e),
-               pvalueAdj = p.adjust(pvalue, method = adjust)) %>% 
-        bind_cols(group_means) %>% 
-        dplyr::select(feature, pvalue, pvalueAdj, everything()) %>% 
+        dplyr::rename_all(~ paste0("pvalue_", .)) %>% 
+        dplyr::mutate(across(starts_with("pvalue"), ~ p.adjust(., method = adjust), .names = "pvalueAdj_{.col}"),
+                      feature = colnames(e)) %>% 
+        dplyr::rename_at(vars(starts_with("pvalueAdj_pvalue_")), ~ gsub("pvalueAdj_pvalue", "pvalueAdj", .)) %>% 
+        bind_cols(group_means, group_sd) %>% 
+        dplyr::select(feature, everything()) %>% 
         dplyr::as_tibble()
 
       return(res_aov_cov)
@@ -162,7 +176,7 @@ PomaUnivariate <- function(data,
                                                                         paired = paired)$p.value})) %>% 
       rownames_to_column("feature") %>%
       mutate(pvalueAdj = p.adjust(pvalue, method = adjust)) %>%
-      bind_cols(group_means) %>%
+      bind_cols(group_means, group_sd) %>%
       mutate(FC = as.numeric(round(group_means[,2]/group_means[,1], 3)),
              diff_means = as.numeric(round(group_means[,2] - group_means[,1], 3))) %>% 
       dplyr::select(feature, FC, diff_means, pvalue, pvalueAdj, everything()) %>% 
@@ -176,7 +190,7 @@ PomaUnivariate <- function(data,
     res_kruskal <- data.frame(pvalue = apply(e, 2, function(x){kruskal.test(x ~ as.factor(Group))$p.value})) %>%
       mutate(pvalueAdj = p.adjust(pvalue, method = adjust),
              kw_rank_sum = apply(e, 2, function(x){kruskal.test(x ~ as.factor(Group))$statistic})) %>% 
-      bind_cols(group_means) %>% 
+      bind_cols(group_means, group_sd) %>% 
       rownames_to_column("feature") %>%
       dplyr::select(feature, kw_rank_sum, pvalue, pvalueAdj, everything()) %>% 
       dplyr::as_tibble()