+++ R/zzz.R
@@ -3,7 +3,7 @@
   suppressWarnings({
     
     poma_welcome <- "Welcome to POMA!"
-    ver <- paste(utils::packageVersion("POMA"), "version")
+    ver <- paste("Version", utils::packageVersion("POMA"))
     poma_shiny <- "POMAShiny app: https://github.com/pcastellanoescuder/POMAShiny"
     info <- "For more detailed package information please visit https://pcastellanoescuder.github.io/POMA/"